requirements.txt
torch>=2.0.0
torchvision==0.16.1
torchaudio==2.1.1
transformers>=4.30.0
datasets>=2.12.0
peft>=0.4.0
sentence-transformers==2.2.2
scikit-learn>=1.2.2
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
accelerate==1.6.0
pyyaml>=6.0
pytest>=7.3.1
black==23.7.0
isort==5.12.0
flake8==6.1.0
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking
seaborn>=0.12.0   # For visualization 